+++ src/grpc_gateway/ibc/applications/interchain_accounts/controller/v1/query.pb.ts
@@ -6,6 +6,15 @@
 
 import * as fm from "../../../../../fetch.pb"
 import * as IbcApplicationsInterchain_accountsControllerV1Controller from "./controller.pb"
+export type QueryInterchainAccountRequest = {
+  owner?: string
+  connection_id?: string
+}
+
+export type QueryInterchainAccountResponse = {
+  address?: string
+}
+
 export type QueryParamsRequest = {
 }
 
@@ -14,6 +23,9 @@ export type QueryParamsResponse = {
 }
 
 export class Query {
+  static InterchainAccount(req: QueryInterchainAccountRequest, initReq?: fm.InitReq): Promise<QueryInterchainAccountResponse> {
+    return fm.fetchReq<QueryInterchainAccountRequest, QueryInterchainAccountResponse>(`/ibc/apps/interchain_accounts/controller/v1/owners/${req["owner"]}/connections/${req["connection_id"]}?${fm.renderURLSearchParams(req, ["owner", "connection_id"])}`, {...initReq, method: "GET"})
+  }
   static Params(req: QueryParamsRequest, initReq?: fm.InitReq): Promise<QueryParamsResponse> {
     return fm.fetchReq<QueryParamsRequest, QueryParamsResponse>(`/ibc/apps/interchain_accounts/controller/v1/params?${fm.renderURLSearchParams(req, [])}`, {...initReq, method: "GET"})
   }