+++ src/grpc_gateway/ibc/applications/transfer/v1/tx.pb.ts
@@ -15,9 +15,11 @@ export type MsgTransfer = {
   receiver?: string
   timeout_height?: IbcCoreClientV1Client.Height
   timeout_timestamp?: string
+  memo?: string
 }
 
 export type MsgTransferResponse = {
+  sequence?: string
 }
 
 export class Msg {