+++ src/grpc_gateway/ibc/applications/transfer/v2/packet.pb.ts
@@ -8,4 +8,5 @@ export type FungibleTokenPacketData = {
   amount?: string
   sender?: string
   receiver?: string
+  memo?: string
 }