+++ src/protobuf/ibc/applications/transfer/v1/tx.ts
@@ -32,10 +32,15 @@ export interface MsgTransfer {
    * The timeout is disabled when set to 0.
    */
   timeout_timestamp: string;
+  /** optional memo */
+  memo: string;
 }
 
 /** MsgTransferResponse defines the Msg/Transfer response type. */
-export interface MsgTransferResponse {}
+export interface MsgTransferResponse {
+  /** sequence number of the transfer packet sent */
+  sequence: string;
+}
 
 function createBaseMsgTransfer(): MsgTransfer {
   return {
@@ -46,6 +51,7 @@ function createBaseMsgTransfer(): MsgTransfer {
     receiver: "",
     timeout_height: undefined,
     timeout_timestamp: "0",
+    memo: "",
   };
 }
 
@@ -74,6 +80,9 @@ export const MsgTransfer = {
     }
     if (message.timeout_timestamp !== "0") {
       writer.uint32(56).uint64(message.timeout_timestamp);
+    }
+    if (message.memo !== "") {
+      writer.uint32(66).string(message.memo);
     }
     return writer;
   },
@@ -106,6 +115,9 @@ export const MsgTransfer = {
         case 7:
           message.timeout_timestamp = longToString(reader.uint64() as Long);
           break;
+        case 8:
+          message.memo = reader.string();
+          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -129,6 +141,7 @@ export const MsgTransfer = {
       timeout_timestamp: isSet(object.timeout_timestamp)
         ? String(object.timeout_timestamp)
         : "0",
+      memo: isSet(object.memo) ? String(object.memo) : "",
     };
   },
 
@@ -148,6 +161,7 @@ export const MsgTransfer = {
         : undefined);
     message.timeout_timestamp !== undefined &&
       (obj.timeout_timestamp = message.timeout_timestamp);
+    message.memo !== undefined && (obj.memo = message.memo);
     return obj;
   },
 
@@ -168,19 +182,23 @@ export const MsgTransfer = {
         ? Height.fromPartial(object.timeout_height)
         : undefined;
     message.timeout_timestamp = object.timeout_timestamp ?? "0";
+    message.memo = object.memo ?? "";
     return message;
   },
 };
 
 function createBaseMsgTransferResponse(): MsgTransferResponse {
-  return {};
+  return { sequence: "0" };
 }
 
 export const MsgTransferResponse = {
   encode(
-    _: MsgTransferResponse,
+    message: MsgTransferResponse,
     writer: _m0.Writer = _m0.Writer.create(),
   ): _m0.Writer {
+    if (message.sequence !== "0") {
+      writer.uint32(8).uint64(message.sequence);
+    }
     return writer;
   },
 
@@ -191,6 +209,9 @@ export const MsgTransferResponse = {
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
+        case 1:
+          message.sequence = longToString(reader.uint64() as Long);
+          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -199,19 +220,23 @@ export const MsgTransferResponse = {
     return message;
   },
 
-  fromJSON(_: any): MsgTransferResponse {
-    return {};
+  fromJSON(object: any): MsgTransferResponse {
+    return {
+      sequence: isSet(object.sequence) ? String(object.sequence) : "0",
+    };
   },
 
-  toJSON(_: MsgTransferResponse): unknown {
+  toJSON(message: MsgTransferResponse): unknown {
     const obj: any = {};
+    message.sequence !== undefined && (obj.sequence = message.sequence);
     return obj;
   },
 
   fromPartial<I extends Exact<DeepPartial<MsgTransferResponse>, I>>(
-    _: I,
+    object: I,
   ): MsgTransferResponse {
     const message = createBaseMsgTransferResponse();
+    message.sequence = object.sequence ?? "0";
     return message;
   },
 };