+++ src/protobuf/ibc/applications/transfer/v2/packet.ts
@@ -18,10 +18,12 @@ export interface FungibleTokenPacketData {
   sender: string;
   /** the recipient address on the destination chain */
   receiver: string;
+  /** optional memo */
+  memo: string;
 }
 
 function createBaseFungibleTokenPacketData(): FungibleTokenPacketData {
-  return { denom: "", amount: "", sender: "", receiver: "" };
+  return { denom: "", amount: "", sender: "", receiver: "", memo: "" };
 }
 
 export const FungibleTokenPacketData = {
@@ -40,6 +42,9 @@ export const FungibleTokenPacketData = {
     }
     if (message.receiver !== "") {
       writer.uint32(34).string(message.receiver);
+    }
+    if (message.memo !== "") {
+      writer.uint32(42).string(message.memo);
     }
     return writer;
   },
@@ -66,6 +71,9 @@ export const FungibleTokenPacketData = {
         case 4:
           message.receiver = reader.string();
           break;
+        case 5:
+          message.memo = reader.string();
+          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -80,6 +88,7 @@ export const FungibleTokenPacketData = {
       amount: isSet(object.amount) ? String(object.amount) : "",
       sender: isSet(object.sender) ? String(object.sender) : "",
       receiver: isSet(object.receiver) ? String(object.receiver) : "",
+      memo: isSet(object.memo) ? String(object.memo) : "",
     };
   },
 
@@ -89,6 +98,7 @@ export const FungibleTokenPacketData = {
     message.amount !== undefined && (obj.amount = message.amount);
     message.sender !== undefined && (obj.sender = message.sender);
     message.receiver !== undefined && (obj.receiver = message.receiver);
+    message.memo !== undefined && (obj.memo = message.memo);
     return obj;
   },
 
@@ -100,6 +110,7 @@ export const FungibleTokenPacketData = {
     message.amount = object.amount ?? "";
     message.sender = object.sender ?? "";
     message.receiver = object.receiver ?? "";
+    message.memo = object.memo ?? "";
     return message;
   },
 };